+++ packages/color-tag/index.ts
@@ -1,3 +1,3 @@
 import "@life-uikit/global-css/tailwind-css/input.css";
 
-export { ColorTag } from "./src/color-tag";
+export { ColorTag } from "./src";