+++ packages/color-tag/vite.config.ts
@@ -1,29 +1,32 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import dts from 'vite-plugin-dts';
+import dts from "vite-plugin-dts";
+import svgr from "vite-plugin-svgr";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   build: {
     lib: {
-      entry: 'index.ts',
-      name: 'ColorTag',
-      fileName: (format) => `color-tag.${format}.js`
+      entry: "index.ts",
+      name: "ColorTag",
+      fileName: (format) => `color-tag.${format}.js`,
     },
     rollupOptions: {
-      external: ['react', 'react-dom'],
+      external: ["react", "@life-uikit/context"],
       output: {
         globals: {
-          react: 'React',
-          'react-dom': 'ReactDOM',
+          react: "React",
+          "react-dom": "ReactDOM",
         },
       },
     },
   },
-  plugins: [react(),
-  dts({
-    // insertTypesEntry: true,
-  }),],
-  resolve: {
-  },
+  plugins: [
+    react(),
+    dts({
+      // insertTypesEntry: true,
+    }),
+    svgr({ exportAsDefault: true })
+  ],
+  resolve: {},
 });