+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,81 @@
 # @life-uikit/storybook
 
+## 0.0.9
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/color-tag@1.1.7
+  - @life-uikit/fluentui@1.1.7
+  - @life-uikit/context@1.1.7
+  - @life-uikit/antd@2.1.7
+
+## 0.0.8
+
+### Patch Changes
+
+- update interface define
+- Updated dependencies
+  - @life-uikit/color-tag@1.1.6
+  - @life-uikit/fluentui@1.1.6
+  - @life-uikit/context@1.1.6
+  - @life-uikit/antd@2.1.6
+
+## 0.0.7
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/color-tag@1.1.5
+  - @life-uikit/fluentui@1.1.5
+  - @life-uikit/context@1.1.5
+  - @life-uikit/antd@2.1.5
+
+## 0.0.6
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/color-tag@1.1.4
+  - @life-uikit/fluentui@1.1.4
+  - @life-uikit/context@1.1.4
+  - @life-uikit/antd@2.1.4
+
+## 0.0.5
+
+### Patch Changes
+
+- update dist
+- Updated dependencies
+  - @life-uikit/color-tag@1.1.3
+  - @life-uikit/fluentui@1.1.3
+  - @life-uikit/context@1.1.3
+  - @life-uikit/antd@2.1.3
+
+## 0.0.4
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/color-tag@1.1.2
+  - @life-uikit/fluentui@1.1.2
+  - @life-uikit/context@1.1.2
+  - @life-uikit/antd@2.1.2
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @life-uikit/context@1.1.1
+  - @life-uikit/antd@2.1.1
+  - @life-uikit/color-tag@1.1.1
+  - @life-uikit/fluentui@1.1.1
+
 ## 0.0.2
 
 ### Patch Changes