+++ packages/fluentui/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @life-uikit/fluentui
 
+## 1.1.7
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/context@1.1.7
+
+## 1.1.6
+
+### Patch Changes
+
+- update interface define
+- Updated dependencies
+  - @life-uikit/context@1.1.6
+
+## 1.1.5
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/context@1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/context@1.1.4
+
+## 1.1.3
+
+### Patch Changes
+
+- update dist
+- Updated dependencies
+  - @life-uikit/context@1.1.3
+
+## 1.1.2
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @life-uikit/context@1.1.2
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @life-uikit/context@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes